+++ package.json
@@ -13,6 +13,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "react-router-dom": "^6.11.2",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
