+++ src/pages/AboutUs.jsx
@@ -1,0 +1,5 @@
+function AboutUs() {
+    return <section>This is Us</section>;
+  }
+
+  export default AboutUs;
