+++ src/pages/Home.jsx
@@ -1,0 +1,11 @@
+function Home() {
+    return (
+        <>
+            <h1>Página home</h1>
+            <p>Bienvenidos a la app my task list</p>
+        </>
+       
+    );
+  }
+  
+  export default Home;
